Fixed bugs and added file check at startup

+++ scripts/TestLocus.py
@@ -90,7 +90,8 @@ class TestLocus:
         # This isn't a very clean way to purge shared temporary files, 
         # though technically it should usually work? Fix this later
         if "finemap" in self.settings["methods"] or "ecaviar" in self.settings["methods"] or "caviarbf" in self.settings["methods"]:
-            finemap.purge_tmp_files(self)
+            if not self.settings["debug"]:
+                finemap.purge_tmp_files(self)
         
         if "coloc" in self.settings["methods"]:
             try:
